backend/requirements.txt
fastapi==0.115.11
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
pydantic==2.10.6
email-validator==2.2.0
python-jose==3.3.0
passlib==1.7.4
google-genai==1.51.0
psycopg2-binary==2.9.9
